requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
colorama==0.4.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
PyJWT==2.6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tzdata==2023.3
